# public/openapi.yaml
openapi: 3.0.0
info:
  title: 'Swagger dumbel'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/'
    description: 'Api localhost'
paths:
  /api/v1/account:
    get:
      tags:
        - Account
      description: 'Return account.'
      operationId: 744264bb63b5fc5a9aa45a0a6e5c7d7b
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseAccount' }] } }, type: object }
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Account
      description: 'Update account.'
      operationId: 93b2e569b9d7e5238fc05f96ef97d7e9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAccountUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseAccount' }] } }, type: object }
      security:
        -
          bearerAuth: []
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      description: 'Return a bearer token.'
      operationId: fb66b153966d9fac2e999136ef0ade67
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestLogin'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseLogin' }] } }, type: object }
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      description: 'Return a bearer token.'
      operationId: fb835c97347e657e2a627790c4c5817a
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestRegister'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseRegister' }] } }, type: object }
  /api/v1/gym/create:
    post:
      tags:
        - Gym
      description: 'Create gym.'
      operationId: a6c4cb857008453f3aba2e6101b98f45
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGymCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseGym' }] } }, type: object }
      security:
        -
          bearerAuth: []
  '/api/v1/gym/{id}/update':
    put:
      tags:
        - Gym
      description: 'Update gym.'
      operationId: cebe2818a13f09f3874e2092674b877d
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: int
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGymUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: object, allOf: [{ $ref: '#/components/schemas/ResponseGym' }] } }, type: object }
      security:
        -
          bearerAuth: []
  /api/v1/gym/list/own:
    get:
      tags:
        - Gym
      description: 'List gym.'
      operationId: dffc1ee871f055c4e95ee74a388c4aaf
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - { properties: { data: { type: array, items: { $ref: '#/components/schemas/ResponseGym' } } }, type: object }
      security:
        -
          bearerAuth: []
components:
  schemas:
    RequestAccountUpdate:
      properties:
        name:
          type: string
        lastname:
          type: string
        phone:
          type: string
        email:
          type: string
        birthday:
          type: string
        password:
          type: string
      type: object
      xml:
        name: RequestAccountUpdate
    RequestLogin:
      properties:
        phone:
          type: string
        password:
          type: string
        device:
          type: string
      type: object
      xml:
        name: RequestLogin
    RequestRegister:
      properties:
        name:
          type: string
        lastname:
          type: string
        phone:
          type: string
        email:
          type: string
        birthday:
          type: string
        password:
          type: string
        device:
          type: string
        uuid:
          type: string
      type: object
      xml:
        name: RequestRegister
    RequestGymCreate:
      properties:
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
      type: object
      xml:
        name: RequestGymCreate
    RequestGymUpdate:
      properties:
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
      type: object
      xml:
        name: RequestGymUpdate
    ResponseAccount:
      properties:
        id:
          type: integer
        name:
          type: string
        lastname:
          type: string
        phone:
          type: string
        email:
          type: string
        birthday:
          type: string
        uuid:
          type: string
      type: object
      xml:
        name: ResponseAccount
    ResponseLogin:
      properties:
        token:
          type: string
      type: object
      xml:
        name: Response
    ResponseRegister:
      properties:
        token:
          type: string
      type: object
      xml:
        name: ResponseRegister
    ResponseGym:
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      type: object
      xml:
        name: ResponseGym
    Response:
      properties:
        success:
          type: boolean
        message:
          type: string
        meta:
          type: object
        data:
          type: object
      type: object
      xml:
        name: Response
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer